api_gateway/requirements.txt
fastapi
uvicorn
httpx
pyjwt
python-jose
pydantic-settings
starlette
sqlalchemy
mysql-connector-python
python-multipart